+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Changed
+- When outputting a tree in Newick, if the root has only one child,
+  then the root is removed from the output.
 
 ## [1.3.2] -- 2019-12-31
 ### Fixed